feat: change Client to be an interface
In order to help with testability and to also decrease relying on type
specific implementations, changed Client to be an interface.

We also stop bundling a version of the DownloadService within the Client
itself, and instead let the user choose which services they want to use
the client for. Additionally, this didn't make much sense, since a
DownloadService requires a client to begin with.

Finally, because DownloadService is no longer bundled within the Client,
we make Client a publicly accessible data structure on the
DownloadService so that it can be constructed and added outside of the
rtorrent package.